refactor service: int32 to int64 + improved docker and docker compose

+++ db/migrations/inventory/20260404170815_init.sql
@@ -2,7 +2,7 @@
 -- +goose StatementBegin
 CREATE TABLE
    warehouses (
-      id SERIAL PRIMARY KEY,
+      id BIGSERIAL PRIMARY KEY,
       name TEXT NOT NULL,
       location TEXT NOT NULL,
       created_at TIMESTAMPTZ DEFAULT now () NOT NULL
@@ -10,10 +10,10 @@ CREATE TABLE
 
 CREATE TABLE
    products (
-      id SERIAL PRIMARY KEY,
-      warehouse_id INTEGER REFERENCES warehouses (id) NOT NULL,
+      id BIGSERIAL PRIMARY KEY,
+      warehouse_id BIGINT REFERENCES warehouses (id) NOT NULL,
       name TEXT NOT NULL,
-      quantity INTEGER DEFAULT 0 NOT NULL CHECK (quantity >= 0),
+      quantity BIGINT DEFAULT 0 NOT NULL CHECK (quantity >= 0),
       created_at TIMESTAMPTZ DEFAULT now () NOT NULL
    );
